125 reorganize api endpoints mt (#127)
* feat: got rid of the '/webpages'

* feat: reorganized the api endpoints

* feat: fixed jinja test

---------
