+++ darkmoon/api/templates/hash.html
@@ -33,16 +33,16 @@
         <nav>
 
             <ul>
-                <button class="dropbtn"><a href="/webpages">Home</a></button>
+                <button class="dropbtn"><a href="/">Home</a></button>
                 <div class="dropdown">
 
                     <button class="button button1">Tools</button>
 
                     <div class="dropdown-content">
 
-                        <a href="/webpages/hash-lookup">Hash Lookup</a>
+                        <a href="/hash-lookup">Hash Lookup</a>
 
-                        <a href="/webpages/upload">ISO Upload</a>
+                        <a href="/upload">ISO Upload</a>
 
                     </div>
